♻️ refactor: 이미지 업로드 API 공통 분리
